refactor: kakao API 관련 클래스 그룹화
