+++ src/main/java/gift/auth/dto/AuthResponse.java
@@ -6,4 +6,5 @@ public record AuthResponse(
     public static AuthResponse of(String token) {
         return new AuthResponse(token);
     }
+
 }
