+++ src/main/java/gift/auth/util/CustomPasswordGenerator.java
@@ -3,7 +3,7 @@ package gift.auth.util;
 import java.security.SecureRandom;
 import java.util.Base64;
 
-public class KakaoAuthUtil {
+public class CustomPasswordGenerator {
 
     public static String generateTemporaryPassword() {
         SecureRandom random = new SecureRandom();
