+++ src/main/java/gift/kakao/auth/dto/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package gift.auth.dto;
+package gift.kakao.auth.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
