+++ src/main/java/gift/kakao/message/dto/KakaoMessageResponse.java
@@ -1,4 +1,4 @@
-package gift.auth.dto;
+package gift.kakao.message.dto;
 
 public record KakaoMessageResponse(
         Integer result_code
