+++ src/main/java/gift/member/application/MemberService.java
@@ -1,12 +1,9 @@
 package gift.member.application;
 
-import gift.auth.application.KakaoClient;
-import gift.auth.dto.AuthResponse;
-import gift.auth.dto.KakaoTokenResponse;
-import gift.auth.util.KakaoAuthUtil;
 import gift.global.error.CustomException;
 import gift.global.error.ErrorCode;
-import gift.global.security.JwtUtil;
+import gift.kakao.auth.dto.KakaoTokenResponse;
+import gift.kakao.client.KakaoClient;
 import gift.member.dao.MemberRepository;
 import gift.member.dto.MemberDto;
 import gift.member.entity.Member;
