+++ src/test/java/gift/application/OrderServiceTest.java
@@ -1,6 +1,6 @@
 package gift.application;
 
-import gift.auth.application.KakaoClient;
+import gift.kakao.client.KakaoClient;
 import gift.member.application.MemberService;
 import gift.member.entity.Member;
 import gift.order.application.OrderService;
